requirements.txt
openai>=1.30.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
